Implement role_0_stake_addresses for both registration chain and cip509
